Fetch parameter values on connection
When we connect to a device fetch the current value
of every parameter. Store this in the model and allow
the model to notify interested parties that a parameter
has changed. We store values based on parameter
names which are guaranteed to be unique and are
easy to look up if more information is required.

The ParameterModel and SpotValueModel are
updated to have a parameter_changed method which
updates the internal item that corresponds to the
value. Both models just convert the value to a floating
point string for now.

Tests:
 - Connect to an inverter and vcu and verify the
    numbers are updated and make sense.
 - Connect repeatedly and verify the internal item
    caches are cleared and refilled