Format the display of parameters and spot values
To allow parameters and spot values to be formatted
(and eventually edited) this creates a custom item type
that can be stored in the ParameterModel and
SpotValueModel. The custom item contains the
floating point value and a link to the corresponding
OIVariable. This allows the current value to be
rendered correctly as a string.

Editing of parameters just changes the string and does
nothing else at this point.

Tests:
 - Fetch from inverter and VCU and review display
   matches "oic dumpall" output.