Add the ability to list errors on a device
Use a newly defined SDO interface to list error
numbers and times on a device. The error
numbers are mapped to names using the
"lasterr" spot value if possible.

The time is presented in wall-clock time.
Provision is made to allow devices which use
the tick durations different from the 10ms
default (stm32-sine and focci use this). This
will allow the Zombieverter VCU to have the
correct time as it uses a 1 second tick.

A full set of unit tests is provided. A problem
with the creation of OpenInverterNode
instances in tests was fixed. Now a valid and
clean canopen.ObjectDictionary instance is
created for each test.

Tests:
 - Verify that a VCU reports "Command or
    parameter not supported"
 - Verify list of errors on Tesla M3 with stm32-
    sine firmware matches web interface
 - Trigger a new error and verify list is updated