GUI: Add error log support
Add an Error Log tab with columns for the error time and message and a Refresh
button. Fill the list of errors from the OpenInverterNode instance.

The error log is filled on initial connection in the same way as the
parameters.

The error log and parameter fetch are now carried out synchronously which
makes them faster and avoids them clashing with each other as simultaneous
use of an OpenInverterNode SDO client is not allowed.

Tests:
 - Verify device without error log support reports that in the log view
   and in the error list
 - Verify a device with no errors reports that in the list
 - Verify a device with errors displays on connection
 - Veirfy an extending list of errors is reflected when the refresh button is
   pressed
 - Run above tests on Windows 11 and Linux