Allow editing of parameters
Create a delegate that provides type-specific editors for parameters in
the parameter view. Numeric parameters use a spin box, enum parameters
use a dropdown, and bitfield parameters use a custom editor with
checkboxes for each bit.

The limits of the numeric editor is fixed for now but could be
improved later by using parameter metadata.

The model is updated but the parameter is not passed to the device yet.

Tests:
 - Verify behaviour of each editor type