Estimate step size and decimal places for parameters
Use some simple heuristics to set the step size and number of decimal
places for parameters based on their range and unit. Parameters with
units starting with "dig" are treated as integer digit parameters.

This also fixes a bug where the minimum and maximum values were not
converted from fixed-point to floating-point before being used to set
the range of the spin box editor.

Tests:
 - Verify that digit parameters have a step size of 1 and no decimal
   places.
 - Verify a selection of floating point parameters have appropriate step
   sizes based on their ranges.
 - Verify the range is applied correctly in the editor.